+++ dtslint/ts3.5/ReadonlyArray.ts
@@ -1,7 +1,8 @@
 import * as _ from '../../src/ReadonlyArray'
-import { pipe } from '../../src/function'
+import { identity, pipe } from '../../src/function'
 import * as N from '../../src/number'
 import { Ord } from '../../src/Ord'
+import * as E from '../../src/Either'
 
 declare const rus: ReadonlyArray<unknown>
 declare const rns: ReadonlyArray<number>
@@ -165,4 +166,17 @@ pipe(
   _.Do,
   _.bind('a', () => _.of(1)),
   _.bind('b', () => _.of('b'))
+)
+
+//
+// some
+//
+
+// $ExpectType Either<readonly number[], ReadonlyNonEmptyArray<number>>
+pipe(
+  rns,
+  E.fromPredicate(
+    _.some((n) => n > 0),
+    identity
+  )
 )